requirements.txt
langchain
langchain_community
langchain_groq
llama_parse
python-dotenv
fastapi 
uvicorn
python-multipart
